+++ packages/core/src/runtime/context.js
@@ -1,6 +1,5 @@
 // @flow
-import type { Action, Cancel, Cont, Step } from '../types'
-import { Coroutine } from './coroutine'
+import type { Cancel, Step } from '../types'
 
 export class Scope<H> implements Cancel {
   handlers: H
@@ -11,10 +10,10 @@ export class Scope<H> implements Cancel {
     this.cancelers = cancelers
   }
 
-  cancel () {
+  cancel (): void {
     const cancelers = this.cancelers
     this.cancelers = []
-    cancelers.forEach(c => c.cancel())
+    return cancelers.forEach(c => c.cancel())
   }
 }
 
@@ -32,11 +31,4 @@ export const childScopeWith = <H0, H1> (handlers: H1, scope: Scope<H0>): Scope<{
   const child = createScope({ ...scope.handlers, ...handlers })
   scope.cancelers.push(child)
   return child
-}
-
-export const runAction = <H, E, A> (cont: Cont<A>, action: Action<E, A>, scope: Scope<H>): Cancel => {
-  const co = new Coroutine(cont, scope, action)
-  scope.cancelers.push(co)
-  co.run()
-  return co
 }
