+++ packages/core/src/runtime/index.js
@@ -2,4 +2,5 @@
 export * from './cancel'
 export * from './context'
 export * from './coroutine'
+export * from './handle'
 export * from './result'
